requirements.txt
Flask>=2.0
flask-cors>=3.0
google-generative-ai>=0.3.0
PyMuPDF>=1.22.5
gTTS>=2.3.0
python-docx>=0.8.11
python-dotenv>=1.0.0
